third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/multiple-host-pseudos-in-compound-expected.txt

CONSOLE ERROR: line 7: Uncaught TypeError: host.createShadowRoot is not a function
Multiple :host(-context) pseudos in same compound.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE

Should be green
